sweetalert + seleccionar y mostrar una imagen en local

+++ src/modules/daybook/views/EntryView.vue
@@ -10,6 +10,9 @@
             </div>
 
             <div>
+
+                <input type="file" @change="onSelectedImage">
+
                 <button v-if="entry.id" class="btn btn-danger mx-2" @click="onDeleteEntry">
                     Borrar
                     <i class="fa fa-trash-alt"></i>
@@ -31,8 +34,13 @@
         </div>
 
 
-        <img 
+        <!-- <img 
             src="https://www.robertlandscapes.com/wp-content/uploads/2014/11/landscape-322100_1280.jpg" 
+            alt="entry-picture"
+            class="img-thumbnail"> -->
+        <img 
+            v-if="localImage"
+            :src="localImage" 
             alt="entry-picture"
             class="img-thumbnail">
 
@@ -48,6 +56,7 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapActions, mapGetters } from 'vuex' // computed!!!
+import Swal from 'sweetalert2'
 
 import getDayMonthYear from '../helpers/getDayMonthYear'
 
@@ -64,7 +73,9 @@ export default {
 
     data() {
         return {
-            entry: null
+            entry: null,
+            localImage: null,
+            file: null
         }
     },
 
@@ -104,6 +115,12 @@ export default {
             this.entry = entry
         },
         async saveEntry(){
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
+
             if(this.entry.id){
                 //actualizar
                 await this.updateEntry(this.entry)
@@ -114,13 +131,46 @@ export default {
                 this.$router.push({ name: 'entry', params: { id }})
             }
 
+            Swal.fire('Guardado', 'Entrada registrada con exito!', 'success')
+
         },
         async onDeleteEntry(){
-            console.log('delete', this.entry)
+            const { isConfirmed } = await Swal.fire({
+                title: '¿Esta seguro?',
+                text: 'Una vez borrado, no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText: 'Si, estoy seguro'
+            })
 
-            await this.deleteEntry(this.entry.id)
+            if( isConfirmed ){
+                new Swal({
+                    title: 'Espere por favor...',
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+                await this.deleteEntry(this.entry.id)
+                this.$router.push({ name: 'no-entry' })
 
-            this.$router.push({ name: 'no-entry' })
+                Swal.fire('Eliminado', '', 'success')
+            }
+        },
+        onSelectedImage(event){
+            const file  = event.target.files[0]
+
+            if(!file){
+                this.localImage = null
+                this.file = null
+                return
+            }
+
+            this.file = file
+
+            const fr = new FileReader()
+            fr.onload = () => this.localImage = fr.result
+            fr.readAsDataURL(file)
+        },
+        onSelectImage(){
+            
         }
     },
